Combine all the part (MCU, gateway, app)
For MCU, a simple parser is implemented. The MCU can now send and
receive UART data with specific format. Especially on data that
correspond to light and pump command, the MCU will repeat the received
command to the gateway - this can be see as a simple acknowledgement
mechanism.

For gateway, all the MQTT subscription and connection beside temp and
humidity data has been change to have QoS of 1. When receiving the
repeat data (act like acknowledgement), publish that data to various
`ack` topic.

For app, QoS has also been upgraded to 1. Button to toggle the light and
pump will now only becomes available after knowing the current state of
the light/pump (either as a retained last value or a from a
acknowledgement topic).

Combined, a rudimentary system with 2-hop acknowledgement is achieved.